Added a couple character sprites to test sizing, fixed a UI issue
All the panels will eventually be subbed for real textures. I had to move some objects around on the scene to make sure all the objects could fit - specifically the date sprites and their Affection meters. Also the date panels, had to add another so the text wouldn't mess up the appearance of the sprites.